move zng flattener from zeekio to zng/flattener
This commit moves the flattening logic out of the zeekio package
and into its own package under zng since this operation is a
generic zng transformation and is not specific to zio.

This will be helpful also to the csv writer when we add it.
