+++ zng/flattener/flattener.go
@@ -1,4 +1,4 @@
-package zeekio
+package flattener
 
 import (
 	"fmt"
@@ -17,7 +17,7 @@ type Flattener struct {
 // records where the type context of the received records must match the
 // zctx parameter provided here.  Any new type descriptors that are created
 // to flatten types also use zctx.
-func NewFlattener(zctx *resolver.Context) *Flattener {
+func New(zctx *resolver.Context) *Flattener {
 	return &Flattener{
 		zctx: zctx,
 		// This mapper maps types back into the same context and gives
